Allow for commits that aren't attached to a PR.
As reported in #21, if a commit that is not associated with a pull
request was entered into LaunchBar and sent to the GitHub action, it
would fail, since there would be no results back from the GraphQL API
call.

This commit resolves that issue by checking the number of returned
results and opening the appropriate link in the browser.