Set token from clipboard
This resolves a compatibility problem with live feedback and setting
tokens. Unfortunately when typing something like:

!set-token partial-token-string

The regex matches and immediately invokes App#setToken, which sets the
token prematurely.

This commit modifies the token setting logic to grab the input directly
from the user's clipboard (which is convenient since that's typically
how the user would enter it anyway.