Add initial pass at request locks
In some cases, the suggestion script may fire off API request so quickly
that it overwhelms GitHub's rate limit offenses and temporarily blocks
the user from making requests.

This commit is a spike on "request locking" which aims to create a
"lock" for a particular request so that it can't fire off again until
the request is unlocked (by completing).