Improve onboarding with support OAuth login
This commit updates the instructions that appeared on the first-use
alert to instead redirect the user right to
http://launchbar-github.com/install to begin the GitHub OAuth dance.
Once complete, the OAuth callback points us to the application with the
token supplied.

This commit also updates various calls to GraphQL responses to ensure
that we're safely accessing information from an object which may not
have the appropriate keys.